DBZ-3634 Enable DDL events to be filtered by include/exclude lists
